# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Coordinate {
  latitude: number;
  longitude: number;
}

export interface BusStop {
  id: string;
  name: string;
  latitude: number;
  longitude: number;
  sequence: number;
  minutesFromPrevious: number;
}

export interface BusLocation {
  busId: string;
  latitude: number;
  longitude: number;
  nextStopId: string;
  nextStop: string;
  etaMinutes: number;
  status: string;
  updatedAt: string;
  source: string;
}

export interface BusLocationInput {
  busId: string;
  latitude: number;
  longitude: number;
  timestamp?: string;
}

export interface BusOccupancyInput {
  /** @minimum 0 */
  currentOccupancy: number;
}

export interface Bus {
  id: string;
  busNumber: string;
  origin: string;
  destination: string;
  routeLabel: string;
  capacity: number;
  currentOccupancy: number;
  currentLocation: Coordinate;
  nextStop: string;
  nextStopId: string;
  etaMinutes: number;
  status: string;
  updatedAt: string;
}

export interface QueueEntry {
  studentId: string;
  busId: string;
  boardingStop: string;
  queuePosition: number;
  joinedAt: string;
  status: string;
}

export interface QueueStatus {
  joined: boolean;
  /** @nullable */
  entry: QueueEntry | null;
  busId: string;
  currentOccupancy: number;
  capacity: number;
  seatsAvailable: number;
  estimatedAvailabilityMinutes: number;
  message: string;
}

export interface QueueInput {
  studentId: string;
  busId: string;
  boardingStop: string;
}

export interface QueueLeaveInput {
  studentId: string;
  busId: string;
}

export interface Notification {
  id: string;
  type: string;
  title: string;
  message: string;
  createdAt: string;
  read: boolean;
  busId: string;
}

export interface NotificationReadInput {
  id: string;
}

export type BusIdQueryParameter = string;

export type GetQueueStatusParams = {
busId?: BusIdQueryParameter;
};