# lib/api-zod/src/generated/types/busLocation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BusLocation {
  busId: string;
  latitude: number;
  longitude: number;
  nextStopId: string;
  nextStop: string;
  etaMinutes: number;
  status: string;
  updatedAt: Date;
  source: string;
}