# lib/api-zod/src/generated/types/busLocationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BusLocationInput {
  busId: string;
  latitude: number;
  longitude: number;
  timestamp?: Date;
}