# lib/api-zod/src/generated/types/busOccupancyInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BusOccupancyInput {
  /** @minimum 0 */
  currentOccupancy: number;
}