# lib/api-zod/src/generated/types/queueInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QueueInput {
  studentId: string;
  busId: string;
  boardingStop: string;
}