# lib/api-zod/src/generated/types/queueStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QueueEntry } from './queueEntry';

export interface QueueStatus {
  joined: boolean;
  /** @nullable */
  entry: QueueEntry | null;
  busId: string;
  currentOccupancy: number;
  capacity: number;
  seatsAvailable: number;
  estimatedAvailabilityMinutes: number;
  message: string;
}